chatbot/requirements.txt
dotenv
google-genai==1.72.0
fastapi==0.135.3
PyMySQL==1.1.1
qdrant-client==1.17.1
SQLAlchemy==2.0.49
uvicorn==0.44.0
requests==2.32.5
langchain-text-splitters==1.1.1
pdfplumber==0.11.9
